refactor(products): reuse shop overview card on admin details + show end times
The admin product details page rendered a bespoke "key facts" grid that
showed each session's stored duration ("90 min") rather than an end time,
duplicating schedule/location/age/language logic the shop already had.

- Lift renderScheduleLinesForDetail into format-product-schedule.ts so the
  detail-page schedule renderer is shared, not copied in the overview card.
- Rename ProductWhenWhereCard -> ProductOverviewCard (file + symbol) and
  retitle its heading "When & where" -> "Good to know" (key whenWhere ->
  overview), since it now also carries age range and language. Narrow its
  prop to a structural Pick so both ProductBrowseRow and the admin detail
  row satisfy it.
- Admin details now reuses ProductOverviewCard for the shared four facts and
  keeps an operational-only card (capacity/waitlist, registration window,
  billing+prices, topic, Padlet) plus a club term-dates fact — the one date
  surface the parent card omits for recurring clubs.
- Lay the overview card out two-up at sm+ (Schedule|Format, Age|Language),
  stacking to one column on mobile.
- Drop the now-dead scheduleRow / fields.{schedule,location,ageRange,
  spokenLanguage} / online / onlineWithin keys; add fields.termDates.
